+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,73 +1,73 @@
-import { Component } from "@angular/core";
+import { Component, OnInit, inject } from "@angular/core";
 import { CartService } from "../../services/cart.service";
-import { ProductModel } from "src/app/products/models/product.model";
-import { ProductsService } from "src/app/products/services/products.service";
 import { CommonModule } from "@angular/common";
 import { CartItemComponent } from "../cart-item/cart-item.component";
 import { CartItemModel } from "../../models/cart-item.model";
 import { FilterCartItemsByCategoryPipe } from "../../pipes/filter-cart-items-by-category.pipe";
-import { map } from "rxjs";
+import { Observable, mergeMap, take, tap } from "rxjs";
 import { OrderByPipe } from "src/app/shared/pipes/order-by.pipe";
 import { DeepKeyOf } from "src/app/shared/deepkeyof.type";
+import { RouterLink } from "@angular/router";
+import { ProductsService } from "src/app/products/services/products.service";
+import { AppSettingsService } from "src/app/core/services/app-settings.service";
+import { SortOption } from "src/app/core/enums/sort-option.enum";
+import { FormsModule } from "@angular/forms";
 
 @Component({
     selector: "app-cart-list",
     standalone: true,
     templateUrl: "./cart-list.component.html",
-    styleUrls: ["./cart-list.component.css"],
-    imports: [CommonModule, CartItemComponent, FilterCartItemsByCategoryPipe, OrderByPipe]
+    imports: [CommonModule, CartItemComponent, FilterCartItemsByCategoryPipe, OrderByPipe, FormsModule, RouterLink]
 })
-export class CartListComponent {
-    isCartPopupOpen: boolean = false;
-    cartItems$ = this.cartService.getProducts();
-    categories$ = this.cartItems$.pipe(
-        map(cartItems => new Set(cartItems.map(item => item.product.category))));
+export class CartListComponent implements OnInit {
+    private readonly productsService = inject(ProductsService);
+    readonly cartService = inject(CartService);
 
-    quantityMapping: { [key: string]: string } = { "=1": "# pc.", other: "# pcs." };
-    sortOptions = [
-        { name: 'ASC', value: true },
-        { name: 'DESC', value: false },
-    ];
+    cartItems$: Observable<readonly CartItemModel[]> | undefined;
+
+    readonly sortOptions = [SortOption.ASC, SortOption.DESC];
     readonly orderByProps: Array<DeepKeyOf<CartItemModel>> = ["cost", "quantity", "product.name"];
-    orderByAsc: boolean = false;
+    currentSortOrder: SortOption = SortOption.ASC;
 
-    constructor(public readonly cartService: CartService,
-        private readonly productsService: ProductsService) {
+    constructor(private readonly appSettingsService: AppSettingsService) {
+        this.appSettingsService.appSettings$.pipe(
+            take(1),
+            tap(settings => this.currentSortOrder = settings.sortOrder)
+        ).subscribe();
     }
 
-    toggleCart(): void {
-        this.isCartPopupOpen = !this.isCartPopupOpen;
+    ngOnInit(): void {
+        this.cartItems$ = this.cartService.getProducts();
     }
 
-    buyRandom(): void {
-        this.cartService.addProduct(this.productsService.getRandomProduct());
+    trackByFn(_: number, item: CartItemModel): number {
+        return item.id;
     }
 
-    clearCart(): void {
-        this.cartService.removeAllProducts();
-        this.isCartPopupOpen = false;
-    }
-
-    trackProductByName(_: number, item: CartItemModel): string {
-        return item.product.name;
-    }
-
-    onRemoveItem(product: ProductModel): void {
-        this.cartService.removeProduct(product);
-        if (this.cartService.isEmptyCart()) {
-            this.isCartPopupOpen = false;
-        }
+    onRemoveItem(item: CartItemModel): void {
+        this.cartItems$ = this.cartService.removeCartItem(item);
     }
 
     onQuantityChange(cartItem: CartItemModel, newQuantity: number): void {
         if (newQuantity > cartItem.quantity) {
-            this.cartService.increaseQuantity(cartItem.product, newQuantity - cartItem.quantity);
+            this.cartItems$ = this.cartService.increaseQuantity(cartItem, newQuantity - cartItem.quantity);
         } else if (newQuantity < cartItem.quantity) {
-            this.cartService.decreaseQuantity(cartItem.product, cartItem.quantity - newQuantity);
+            this.cartItems$ = this.cartService.decreaseQuantity(cartItem, cartItem.quantity - newQuantity);
         }
     }
 
-    onSortOptionChange(value: string): void {
-        this.orderByAsc = value === "true";
+    onSortOptionChange(value: SortOption): void {
+        this.currentSortOrder = value;
+        this.appSettingsService.updateSettings("sortOrder", value);
+    }
+
+    clearCart(): void {
+        this.cartItems$ = this.cartService.removeAllProducts();
+    }
+
+    buyRandom(): void {
+        this.cartItems$ = this.productsService.getRandomProduct().pipe(
+            mergeMap(item => this.cartService.addProduct(item))
+        );
     }
 }