+++ src/app/core/components/errors/error-404.component.ts
@@ -1,0 +1,12 @@
+import { Component } from "@angular/core";
+
+@Component({
+    selector: "app-error-404",
+    standalone: true,
+    template: `
+    <h1>Error 404 - Not Found</h1>
+    <p>Sorry, the page you are looking for does not exist.</p>
+  `
+})
+export class Error404Component {
+}