+++ src/app/core/models/app-settings.model.ts
@@ -1,0 +1,5 @@
+import { SortOption } from "../enums/sort-option.enum";
+
+export interface AppSettingsModel {
+    sortOrder: SortOption;
+}