+++ src/app/order/components/process-order/process-order.component.ts
@@ -1,0 +1,9 @@
+import { Component } from "@angular/core";
+
+@Component({
+    selector: "app-process-order",
+    standalone: true,
+    templateUrl: "./process-order.component.html"
+})
+export class ProcessOrderComponent {
+}