+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,28 +1,33 @@
-import { Component, OnInit } from "@angular/core";
+import { Component, OnInit, inject } from "@angular/core";
 import { ProductsService } from "../../services/products.service";
 import { ProductModel } from "../../models/product.model";
 import { CartService } from "src/app/cart/services/cart.service";
 import { CommonModule } from "@angular/common";
 import { ProductComponent } from "../product/product.component";
+import { Observable, tap, take } from "rxjs";
 
 @Component({
     selector: "app-product-list",
     standalone: true,
     templateUrl: "./product-list.component.html",
-    styleUrls: ["./product-list.component.css"],
     imports: [CommonModule, ProductComponent]
 })
 export class ProductListComponent implements OnInit {
-    productList!: ProductModel[];
-
-    constructor(private readonly productsService: ProductsService,
-        private readonly cartService: CartService) { }
+    products$: Observable<ProductModel[]> | undefined;
+    loading: boolean = true;
+    private readonly cartService = inject(CartService);
+    private readonly productsService = inject(ProductsService);
 
     ngOnInit(): void {
-        this.productList = this.productsService.getProducts();
+        this.products$ = this.productsService.getProducts().pipe(
+            take(1),
+            tap(() => this.loading = false)
+        );
     }
 
-    onAddToCart(product: ProductModel): void {
-        this.cartService.addProduct(product);
+    onAddToCart(product: ProductModel) {
+        this.cartService.addProduct(product).pipe(
+            take(1)
+        ).subscribe();
     }
 }