+++ src/app/products/models/product.model.ts
@@ -1,6 +1,7 @@
 import { Category } from "../enums/category.enum";
 
 export interface ProductModel {
+    id: number,
     name: string;
     description: string;
     price: number;