+++ Changelog.md
@@ -10,6 +10,56 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## v1.0.6 - 2023-09-16
+
+### Added
+
+- concurrently & json-server [`585295b`](https://github.com/try-once-more/shop/commit/585295b7739207ef75c8ec05b7c6ade665099421)
+- HttpClient, CartObservableService, ProductsPromiseService [`9b323e8`](https://github.com/try-once-more/shop/commit/9b323e8ea31d1e99c7b13f6757b003334fc9c184)
+- AppSettingsService [`f95ccca`](https://github.com/try-once-more/shop/commit/f95cccad698bcd99535e99dd26e5730f234a5842)
+- TimingInterceptor [`b4e55d0`](https://github.com/try-once-more/shop/commit/b4e55d0b1a81372659b65e6b4dcff9f2d2d2bbe3)
+
+### Fixed
+
+- styles [`2405f61`](https://github.com/try-once-more/shop/commit/2405f6131a63ca1dd45d0bb90b2345cc32a883ab)
+- ProductViewComponent [`374d691`](https://github.com/try-once-more/shop/commit/374d69185eab16941a3a63d512a8db0ef9ad4c5b)
+
+### Changed
+
+- cart items and product list now stored in a database [`3253cb7`](https://github.com/try-once-more/shop/commit/3253cb7d6eab97ee60e54d24dfe1e3c03e33cf83)
+- sortOption type [`4c47e03`](https://github.com/try-once-more/shop/commit/4c47e03a97587f06747e6ab0c9a513c80c1f01c3)
+- LocalStorageService became generic [`2f94809`](https://github.com/try-once-more/shop/commit/2f94809bdcaff2683873418667a9fa1c87b59364)
+
+## v1.0.5 - 2023-09-14
+
+### Added
+
+- AdminComponent,  OrdersComponent, ProductViewComponent, ADMIN_ROUTES, isAdminGuard, AuthService, [`237fd8c`](https://github.com/try-once-more/shop/commit/237fd8cc89014bf960b628a387312495daac80f5)
+- cart path. Some functionality moved from CartListComponent to HeaderComponent. [`295cb09`](https://github.com/try-once-more/shop/commit/295cb09883e80b572a99a111e482f0d3b36fbc23)
+- storage\recovery of all products and added to cart in\from localStorage [`4fc2b58`](https://github.com/try-once-more/shop/commit/4fc2b58c36752fadcebe0246d6e40678ea41fb37)
+- productResolver, productTitleResolver and product/:productID path with dynamic title. [`c8d5f52`](https://github.com/try-once-more/shop/commit/c8d5f528e3e54ad0e56a8345eb278e604db70a77)
+- bootstrap v5.3.1 and bootstrap-icons 1.10.5 [`952ccab`](https://github.com/try-once-more/shop/commit/952ccab9f30409d8136a89437c09bdf7cd8fdc6e)
+- canDeactivateGuard [`bb59d34`](https://github.com/try-once-more/shop/commit/bb59d3431a8a93aa569354c18deb313e36b89f04)
+- cart/order path to the router that uses ProcessOrderComponent and isCartEmptyGuard. [`4bae7dd`](https://github.com/try-once-more/shop/commit/4bae7dd818bc7bfdd39852f156521bcda0b9343f)
+- route partitioning by areas [`b0d0d38`](https://github.com/try-once-more/shop/commit/b0d0d380de9edfed8714057ef3c1d531c2b498c3)
+- ProcessOrderComponent [`ab07090`](https://github.com/try-once-more/shop/commit/ab0709065f36cf60a65befd328a18fc95688455c)
+- view and edit buttons to ProductComponent [`c9c0a77`](https://github.com/try-once-more/shop/commit/c9c0a775435acde870876bcb4d95b9cd8fc20391)
+- redirection to 404 and 403 [`cca7d94`](https://github.com/try-once-more/shop/commit/cca7d94f91afbc553ef1244570af5154f9758245)
+- routing, products-list as default path [`946d19a`](https://github.com/try-once-more/shop/commit/946d19a1d1fcccf071203f9f3f637cf07ec8d5bf)
+- isCartEmptyGuard [`f563f28`](https://github.com/try-once-more/shop/commit/f563f28323ac7e80bee0954a277c9fc3022016d7)
+- navigation to admin page after login as admin [`dcf6c43`](https://github.com/try-once-more/shop/commit/dcf6c43daed8cdb3ada9077c9e9ef71c1d4b9054)
+
+### Fixed
+
+- cart-list component and it routing [`94a5c68`](https://github.com/try-once-more/shop/commit/94a5c68b17bde99e0aa79aad3476632ae654e56a)
+- navigation to the previous page if isCartEmptyGuard denied access [`5f1f245`](https://github.com/try-once-more/shop/commit/5f1f245b2a24a43f588e0709f64e14939d41f9c9)
+
+### Changed
+
+- **Breaking change:** application to use bootstrap styles [`ac6d254`](https://github.com/try-once-more/shop/commit/ac6d2548667f59bba18576e42344617edefe0872)
+- ProductViewComponent [`9a9b486`](https://github.com/try-once-more/shop/commit/9a9b48626e27a9e50d65dcb15599e470158da6b7)
+- getProducts to observable [`47cb80f`](https://github.com/try-once-more/shop/commit/47cb80f8940273363dd5e17c8f3a4311e174821b)
+
 ## v1.0.4 - 2023-08-29
 
 ### Added